Moves registerThemingParticipant to view.
